[tests] Restrict the workflow for a tag
This commit adds a rule to restrict the tests workflow
to run when a tag is pushed to the repository.

Whenever the maintainer using the `publish` command,
a release commit and a tag is pushed. This would trigger
the tests workflow twice resulting in redundant workflows
in the logs. This change is needed to avoid this.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>